+++ src/test/java/io/github/astrapi69/xml/ObjectToXmlExtensionsTest.java
@@ -27,11 +27,18 @@ package io.github.astrapi69.xml;
 import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.assertNotNull;
 
+import java.security.KeyPair;
+import java.security.NoSuchAlgorithmException;
+import java.security.NoSuchProviderException;
+import java.security.PrivateKey;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
+import io.github.astrapi69.crypto.algorithm.KeyPairGeneratorAlgorithm;
+import io.github.astrapi69.crypto.factories.KeyPairFactory;
+import io.github.astrapi69.crypto.key.KeySize;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
@@ -79,8 +86,7 @@ public class ObjectToXmlExtensionsTest
 	 * Test method for {@link ObjectToXmlExtensions#toXmlWithXStream(Object)}
 	 */
 	@Test(enabled = true)
-	public void testToXmlWithXStreamObject()
-	{
+	public void testToXmlWithXStreamObject() throws NoSuchAlgorithmException, NoSuchProviderException {
 		String actual;
 		String expected;
 		Person person;
@@ -97,6 +103,13 @@ public class ObjectToXmlExtensionsTest
 			+ "  </person>\n" + "</io.github.astrapi69.test.objects.Employee>";
 		assertNotNull(actual);
 		assertEquals(actual, expected);
+
+		KeyPair keyPair = KeyPairFactory.newKeyPair(KeyPairGeneratorAlgorithm.DIFFIE_HELLMAN,
+				KeySize.KEYSIZE_2048);
+
+		String xmlWithXStream = ObjectToXmlExtensions.toXmlWithXStream(keyPair.getPrivate());
+		PrivateKey privateKey = XmlToObjectExtensions.toObjectWithXStream(xmlWithXStream);
+		assertEquals(privateKey, keyPair.getPrivate());
 	}
 
 	/**